Binding、interfaceを使用しないよう修正
- Switchの動作のBindingをせず、コードビハインドのメソッドで制御するようにした
- IWeatherService経由で表示データ取得していたものを、コードビハインドのメソッドで直接取得するようにした
