--- Finish_CodeBehind/MobileApp/MobileApp/Services/IWeatherService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MobileApp.Services
-{
-    public interface IWeatherService
-    {
-        Task<ObservableCollection<Weather>> GetWeathersAsync();
-    }
-}
